Complete rewriting of the lexer
The previous implementation started off using only two primitives: peek and advance, effectively consuming the supplied reader one character at a time. It quickly got unwieldy with the iterations and evolutions, and since at one time it was modified to read one line at a time (to help error reporting), the original assumption of reading one char at a time fell apart and ended up as a hinderance. It doesn't buy us anything, since the only case where it could be useful would be huge-one-line-programs, which simply cannot happen with Litil, being a white-space sensitive language.

The new implementation takes advantage of the fact that we read one line at a time to greatly simplify the lexing logic.